// src/app/types/pipelines.ts
export interface Pipeline {
  id: number;
  created_at: string | null;
  updated_at: string | null;
  id_contact: string | null;
  full_name: string | null;
  email: string | null;
  phone: string | null;
  opportunity_name: string | null;
  status: string | null;
  pipeline_stage: string | null;
  pipeline_name: string | null;
  location_name: string | null;
  pipeline_id: string | null;
  location_id: string | null;
  date_created: string | null;
}